+++ src/tensor/buffer.h
@@ -18,11 +18,11 @@ typedef struct buffer_t
     datatype_t datatype;
     void *data;
     size_t size;
-    uint32_t n;
+    uint64_t n;
     bool_t copy;
 } buffer_t;
 
-nw_error_t *buffer_create(buffer_t **buffer, runtime_t runtime, datatype_t datatype, view_t *view, void *data, uint32_t n, bool_t copy);
+nw_error_t *buffer_create(buffer_t **buffer, runtime_t runtime, datatype_t datatype, view_t *view, void *data, uint64_t n, bool_t copy);
 void buffer_destroy(buffer_t *buffer);
 
 nw_error_t *runtime_malloc(buffer_t *buffer);
@@ -48,7 +48,7 @@ nw_error_t *runtime_power(buffer_t *x_buffer, buffer_t *y_buffer, buffer_t *z_bu
 nw_error_t *runtime_compare_equal(buffer_t *x_buffer, buffer_t *y_buffer, buffer_t *z_buffer);
 nw_error_t *runtime_compare_greater(buffer_t *x_buffer, buffer_t *y_buffer, buffer_t *z_buffer);
 nw_error_t *runtime_matrix_multiplication(buffer_t *x_buffer, buffer_t *y_buffer, buffer_t *z_buffer);
-nw_error_t *runtime_summation(buffer_t *x, buffer_t *result, uint32_t axis);
-nw_error_t *runtime_maximum(buffer_t *x, buffer_t *result, uint32_t axis);
+nw_error_t *runtime_summation(buffer_t *x, buffer_t *result, uint64_t axis);
+nw_error_t *runtime_maximum(buffer_t *x, buffer_t *result, uint64_t axis);
 
 #endif
